backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Async HTTP client for JIRA API
httpx==0.26.0

# Configuration
pyyaml==6.0.1

# Database/Cache
aiosqlite==0.19.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Date/time handling
python-dateutil==2.8.2

# CORS support (included in fastapi)
